chore: split PreCheckSecurityOwnershipValidationHook per type
using the new Validator
