+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/validation/DefaultValidators.java
@@ -42,13 +42,16 @@ import org.hisp.dhis.tracker.validation.hooks.EnrollmentDateValidator;
 import org.hisp.dhis.tracker.validation.hooks.EnrollmentGeoValidator;
 import org.hisp.dhis.tracker.validation.hooks.EnrollmentInExistingValidator;
 import org.hisp.dhis.tracker.validation.hooks.EnrollmentNoteValidator;
+import org.hisp.dhis.tracker.validation.hooks.EnrollmentPreCheckSecurityOwnershipValidator;
 import org.hisp.dhis.tracker.validation.hooks.EventCategoryOptValidator;
 import org.hisp.dhis.tracker.validation.hooks.EventDataValuesValidator;
 import org.hisp.dhis.tracker.validation.hooks.EventDateValidator;
 import org.hisp.dhis.tracker.validation.hooks.EventGeoValidator;
 import org.hisp.dhis.tracker.validation.hooks.EventNoteValidator;
+import org.hisp.dhis.tracker.validation.hooks.EventPreCheckSecurityOwnershipValidator;
 import org.hisp.dhis.tracker.validation.hooks.RelationshipsValidator;
 import org.hisp.dhis.tracker.validation.hooks.TrackedEntityAttributeValidator;
+import org.hisp.dhis.tracker.validation.hooks.TrackedEntityPreCheckSecurityOwnershipValidator;
 import org.springframework.stereotype.Component;
 
 /**
@@ -62,6 +65,10 @@ public class DefaultValidators implements Validators
 
     private final TrackedEntityAttributeValidator attributeValidator;
 
+    private final TrackedEntityPreCheckSecurityOwnershipValidator trackedEntityPreCheckSecurityOwnershipValidator;
+
+    private final EnrollmentPreCheckSecurityOwnershipValidator enrollmentPreCheckSecurityOwnershipValidator;
+
     private final EnrollmentNoteValidator enrollmentNoteValidator;
 
     private final EnrollmentInExistingValidator enrollmentInExistingValidator;
@@ -71,6 +78,8 @@ public class DefaultValidators implements Validators
     private final EnrollmentDateValidator enrollmentDateValidator;
 
     private final EnrollmentAttributeValidator enrollmentAttributeValidator;
+
+    private final EventPreCheckSecurityOwnershipValidator eventPreCheckSecurityOwnershipValidator;
 
     private final EventCategoryOptValidator eventCategoryOptValidator;
 
@@ -90,6 +99,7 @@ public class DefaultValidators implements Validators
     public List<Validator<TrackedEntity>> getTrackedEntityValidators()
     {
         return List.of(
+            trackedEntityPreCheckSecurityOwnershipValidator,
             attributeValidator );
     }
 
@@ -97,6 +107,7 @@ public class DefaultValidators implements Validators
     public List<Validator<Enrollment>> getEnrollmentValidators()
     {
         return List.of(
+            enrollmentPreCheckSecurityOwnershipValidator,
             enrollmentNoteValidator,
             enrollmentInExistingValidator,
             enrollmentGeoValidator,
@@ -108,6 +119,7 @@ public class DefaultValidators implements Validators
     public List<Validator<Event>> getEventValidators()
     {
         return List.of(
+            eventPreCheckSecurityOwnershipValidator,
             eventCategoryOptValidator,
             eventDateValidator,
             eventGeoValidator,
