+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/validation/Validator.java
@@ -47,4 +47,9 @@ public interface Validator<T extends TrackerDto>
     {
         return strategy != TrackerImportStrategy.DELETE;
     }
+
+    default boolean skipOnError()
+    {
+        return false;
+    }
 }
