+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/validation/hooks/EventPreCheckSecurityOwnershipValidator.java
@@ -29,9 +29,6 @@ package org.hisp.dhis.tracker.validation.hooks;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 import static org.hisp.dhis.tracker.report.TrackerErrorCode.E1083;
-import static org.hisp.dhis.tracker.report.TrackerErrorCode.E1100;
-import static org.hisp.dhis.tracker.report.TrackerErrorCode.E1103;
-import static org.hisp.dhis.tracker.validation.hooks.TrackerImporterAssertErrors.ENROLLMENT_CANT_BE_NULL;
 import static org.hisp.dhis.tracker.validation.hooks.TrackerImporterAssertErrors.EVENT_CANT_BE_NULL;
 import static org.hisp.dhis.tracker.validation.hooks.TrackerImporterAssertErrors.ORGANISATION_UNIT_CANT_BE_NULL;
 import static org.hisp.dhis.tracker.validation.hooks.TrackerImporterAssertErrors.PROGRAM_CANT_BE_NULL;
@@ -57,22 +54,19 @@ import org.hisp.dhis.program.Program;
 import org.hisp.dhis.program.ProgramInstance;
 import org.hisp.dhis.program.ProgramStage;
 import org.hisp.dhis.program.ProgramStageInstance;
-import org.hisp.dhis.security.Authorities;
 import org.hisp.dhis.security.acl.AclService;
 import org.hisp.dhis.trackedentity.TrackedEntityInstance;
 import org.hisp.dhis.trackedentity.TrackedEntityProgramOwnerOrgUnit;
-import org.hisp.dhis.trackedentity.TrackedEntityType;
 import org.hisp.dhis.trackedentity.TrackerOwnershipManager;
 import org.hisp.dhis.tracker.TrackerImportStrategy;
 import org.hisp.dhis.tracker.bundle.TrackerBundle;
 import org.hisp.dhis.tracker.domain.Enrollment;
 import org.hisp.dhis.tracker.domain.Event;
-import org.hisp.dhis.tracker.domain.TrackedEntity;
 import org.hisp.dhis.tracker.domain.TrackerDto;
 import org.hisp.dhis.tracker.preheat.TrackerPreheat;
 import org.hisp.dhis.tracker.report.TrackerErrorCode;
-import org.hisp.dhis.tracker.validation.TrackerValidationHook;
 import org.hisp.dhis.tracker.validation.ValidationErrorReporter;
+import org.hisp.dhis.tracker.validation.Validator;
 import org.hisp.dhis.user.User;
 import org.springframework.stereotype.Component;
 
@@ -83,8 +77,8 @@ import org.springframework.stereotype.Component;
 @Component
 @RequiredArgsConstructor
 @Slf4j
-public class PreCheckSecurityOwnershipValidationHook
-    implements TrackerValidationHook
+public class EventPreCheckSecurityOwnershipValidator
+    implements Validator<Event>
 {
     @Nonnull
     private final AclService aclService;
@@ -98,152 +92,7 @@ public class PreCheckSecurityOwnershipValidationHook
     private static final String ORG_UNIT_NO_USER_ASSIGNED = " has no organisation unit assigned, so we skip user validation";
 
     @Override
-    public void validateTrackedEntity( ValidationErrorReporter reporter, TrackerBundle bundle,
-        TrackedEntity trackedEntity )
-    {
-        TrackerImportStrategy strategy = bundle.getStrategy( trackedEntity );
-        User user = bundle.getUser();
-
-        checkNotNull( user, USER_CANT_BE_NULL );
-        checkNotNull( trackedEntity, TRACKED_ENTITY_CANT_BE_NULL );
-
-        TrackedEntityType trackedEntityType = strategy.isUpdateOrDelete()
-            ? bundle.getPreheat().getTrackedEntity( trackedEntity.getTrackedEntity() ).getTrackedEntityType()
-            : bundle.getPreheat()
-                .getTrackedEntityType( trackedEntity.getTrackedEntityType() );
-
-        OrganisationUnit organisationUnit = strategy.isUpdateOrDelete()
-            ? bundle.getPreheat().getTrackedEntity( trackedEntity.getTrackedEntity() ).getOrganisationUnit()
-            : bundle.getPreheat().getOrganisationUnit( trackedEntity.getOrgUnit() );
-
-        // If trackedEntity is newly created, or going to be deleted, capture
-        // scope has to be checked
-        if ( strategy.isCreate() || strategy.isDelete() )
-        {
-            checkOrgUnitInCaptureScope( reporter, bundle, trackedEntity, organisationUnit );
-        }
-        // if its to update trackedEntity, search scope has to be checked
-        else
-        {
-            checkOrgUnitInSearchScope( reporter, bundle, trackedEntity, organisationUnit );
-        }
-
-        if ( strategy.isDelete() )
-        {
-            TrackedEntityInstance tei = bundle.getPreheat().getTrackedEntity( trackedEntity.getTrackedEntity() );
-
-            if ( tei.getProgramInstances().stream().anyMatch( pi -> !pi.isDeleted() )
-                && !user.isAuthorized( Authorities.F_TEI_CASCADE_DELETE.getAuthority() ) )
-            {
-                reporter.addError( trackedEntity, E1100, user, tei );
-            }
-        }
-
-        checkTeiTypeWriteAccess( reporter, bundle, trackedEntity, trackedEntityType );
-    }
-
-    private void checkTeiTypeWriteAccess( ValidationErrorReporter reporter, TrackerBundle bundle,
-        TrackedEntity trackedEntity,
-        TrackedEntityType trackedEntityType )
-    {
-        User user = bundle.getUser();
-
-        checkNotNull( user, USER_CANT_BE_NULL );
-        checkNotNull( trackedEntityType, TRACKED_ENTITY_TYPE_CANT_BE_NULL );
-
-        if ( !aclService.canDataWrite( user, trackedEntityType ) )
-        {
-            reporter.addError( trackedEntity, TrackerErrorCode.E1001, user, trackedEntityType );
-        }
-    }
-
-    @Override
-    public void validateEnrollment( ValidationErrorReporter reporter, TrackerBundle bundle, Enrollment enrollment )
-    {
-        TrackerImportStrategy strategy = bundle.getStrategy( enrollment );
-        TrackerPreheat preheat = bundle.getPreheat();
-        User user = bundle.getUser();
-        Program program = strategy.isUpdateOrDelete()
-            ? bundle.getPreheat().getEnrollment( enrollment.getEnrollment() )
-                .getProgram()
-            : bundle.getPreheat().getProgram( enrollment.getProgram() );
-        OrganisationUnit ownerOrgUnit = getOwnerOrganisationUnit( preheat, enrollment.getTrackedEntity(),
-            preheat.getProgram( enrollment.getProgram() ) );
-
-        checkNotNull( user, USER_CANT_BE_NULL );
-        checkNotNull( enrollment, ENROLLMENT_CANT_BE_NULL );
-        checkNotNull( program, PROGRAM_CANT_BE_NULL );
-
-        checkEnrollmentOrgUnit( reporter, bundle, strategy, enrollment, program );
-
-        if ( strategy.isDelete() )
-        {
-            boolean hasNonDeletedEvents = programInstanceHasEvents( preheat, enrollment.getEnrollment() );
-            boolean hasNotCascadeDeleteAuthority = !user
-                .isAuthorized( Authorities.F_ENROLLMENT_CASCADE_DELETE.getAuthority() );
-
-            if ( hasNonDeletedEvents && hasNotCascadeDeleteAuthority )
-            {
-                reporter.addError( enrollment, E1103, user, enrollment.getEnrollment() );
-            }
-        }
-
-        checkWriteEnrollmentAccess( reporter, bundle, enrollment, program, ownerOrgUnit );
-    }
-
-    private OrganisationUnit getOwnerOrganisationUnit( TrackerPreheat preheat, String teiUid, Program program )
-    {
-        Map<String, TrackedEntityProgramOwnerOrgUnit> programOwner = preheat.getProgramOwner()
-            .get( teiUid );
-        if ( programOwner == null || programOwner.get( program.getUid() ) == null )
-        {
-            return null;
-        }
-        else
-        {
-            return programOwner.get( program.getUid() ).getOrganisationUnit();
-        }
-    }
-
-    private boolean programInstanceHasEvents( TrackerPreheat preheat, String programInstanceUid )
-    {
-        return preheat.getProgramInstanceWithOneOrMoreNonDeletedEvent().contains( programInstanceUid );
-    }
-
-    private void checkEnrollmentOrgUnit( ValidationErrorReporter reporter, TrackerBundle bundle,
-        TrackerImportStrategy strategy, Enrollment enrollment, Program program )
-    {
-        OrganisationUnit enrollmentOrgUnit;
-
-        if ( strategy.isUpdateOrDelete() )
-        {
-            enrollmentOrgUnit = bundle.getPreheat().getEnrollment( enrollment.getEnrollment() )
-                .getOrganisationUnit();
-
-            if ( enrollmentOrgUnit == null )
-            {
-                log.warn( "ProgramInstance " + enrollment.getEnrollment()
-                    + ORG_UNIT_NO_USER_ASSIGNED );
-                return;
-            }
-        }
-        else
-        {
-            checkNotNull( enrollment.getOrgUnit().getIdentifierOrAttributeValue(), ORGANISATION_UNIT_CANT_BE_NULL );
-            enrollmentOrgUnit = bundle.getPreheat().getOrganisationUnit( enrollment.getOrgUnit() );
-        }
-
-        // If enrollment is newly created, or going to be deleted, capture scope
-        // has to be checked
-        if ( program.isWithoutRegistration() || strategy.isCreate()
-            || strategy.isDelete() )
-        {
-            checkOrgUnitInCaptureScope( reporter, bundle, enrollment, enrollmentOrgUnit );
-        }
-    }
-
-    @Override
-    public void validateEvent( ValidationErrorReporter reporter, TrackerBundle bundle, Event event )
+    public void validate( ValidationErrorReporter reporter, TrackerBundle bundle, Event event )
     {
         TrackerImportStrategy strategy = bundle.getStrategy( event );
         TrackerPreheat preheat = bundle.getPreheat();
@@ -377,6 +226,20 @@ public class PreCheckSecurityOwnershipValidationHook
         }
     }
 
+    private OrganisationUnit getOwnerOrganisationUnit( TrackerPreheat preheat, String teiUid, Program program )
+    {
+        Map<String, TrackedEntityProgramOwnerOrgUnit> programOwner = preheat.getProgramOwner()
+            .get( teiUid );
+        if ( programOwner == null || programOwner.get( program.getUid() ) == null )
+        {
+            return null;
+        }
+        else
+        {
+            return programOwner.get( program.getUid() ).getOrganisationUnit();
+        }
+    }
+
     @Override
     public boolean needsToRun( TrackerImportStrategy strategy )
     {
@@ -403,20 +266,6 @@ public class PreCheckSecurityOwnershipValidationHook
         }
     }
 
-    private void checkOrgUnitInSearchScope( ValidationErrorReporter reporter, TrackerBundle bundle, TrackerDto dto,
-        OrganisationUnit orgUnit )
-    {
-        User user = bundle.getUser();
-
-        checkNotNull( user, USER_CANT_BE_NULL );
-        checkNotNull( orgUnit, ORGANISATION_UNIT_CANT_BE_NULL );
-
-        if ( !organisationUnitService.isInUserSearchHierarchyCached( user, orgUnit ) )
-        {
-            reporter.addError( dto, TrackerErrorCode.E1003, orgUnit, user );
-        }
-    }
-
     private void checkTeiTypeAndTeiProgramAccess( ValidationErrorReporter reporter, TrackerDto dto,
         User user,
         String trackedEntityInstance,
@@ -438,29 +287,6 @@ public class PreCheckSecurityOwnershipValidationHook
                 program ) )
         {
             reporter.addError( dto, TrackerErrorCode.E1102, user, trackedEntityInstance, program );
-        }
-    }
-
-    private void checkWriteEnrollmentAccess( ValidationErrorReporter reporter, TrackerBundle bundle,
-        Enrollment enrollment, Program program,
-        OrganisationUnit ownerOrgUnit )
-    {
-        User user = bundle.getUser();
-
-        checkNotNull( user, USER_CANT_BE_NULL );
-        checkNotNull( program, PROGRAM_CANT_BE_NULL );
-
-        checkProgramWriteAccess( reporter, enrollment, user, program );
-
-        if ( program.isRegistration() )
-        {
-            String trackedEntity = bundle.getStrategy( enrollment ).isDelete()
-                ? bundle.getPreheat().getEnrollment( enrollment.getEnrollment() ).getEntityInstance().getUid()
-                : enrollment.getTrackedEntity();
-
-            checkNotNull( program.getTrackedEntityType(), TRACKED_ENTITY_TYPE_CANT_BE_NULL );
-            checkTeiTypeAndTeiProgramAccess( reporter, enrollment, user, trackedEntity,
-                ownerOrgUnit, program );
         }
     }
